+++ pyslvs/__init__.py
@@ -2,9 +2,9 @@
 
 """Kernel of Pyslvs."""
 
-__version__ = "20.11.0.dev0"
+__version__ = "21.01.0.dev0"
 __author__ = "Yuan Chang"
-__copyright__ = "Copyright (C) 2016-2020"
+__copyright__ = "Copyright (C) 2016-2021"
 __license__ = "AGPL"
 __email__ = "pyslvs@gmail.com"
 
@@ -12,7 +12,8 @@ from .expression import get_vlinks, VJoint, VPoint, VLink, Coord
 from .bfgs import SolverSystem
 from .topo_config import t_config, EStack
 from .tinycadlib import (
-    pxy, ppp, plap, pllp, plpp, palp, vpoint_dof, expr_solving,
+    pxy, ppp, plap, pllp, plpp, palp, vpoint_dof, expr_solving, uniform_path,
+    uniform_four_bar,
 )
 from .planar_linkage import (
     FMatch, norm_path, curvature, derivative, path_signature, cross_correlation,
@@ -24,9 +25,6 @@ from .expression_parser import (
 from .example import example_list, all_examples
 from .collection import collection_list, all_collections
 from .efd import efd_fitting
-
-# backward alignment
-Coordinate = Coord
 
 
 def get_include() -> str: