+++ pyslvs/planar_linkage.pyx
@@ -5,7 +5,7 @@
 """The callable classes of the validation in the algorithm.
 
 author: Yuan Chang
-copyright: Copyright (C) 2016-2020
+copyright: Copyright (C) 2016-2021
 license: AGPL
 email: pyslvs@gmail.com
 """
@@ -269,6 +269,15 @@ cdef double[:, :] _slice_nan2d(double[:, :] s) nogil:
     if second == -1:
         second = len(s) - 1
     return s[first:second]
+
+
+cdef double trapezoidal_camp(double[:] a, double[:] b):
+    """Error comparison by trapezoidal rule."""
+    cdef double area = 0
+    cdef int i
+    for i in range(1, len(a)):
+        area += abs(a[i - 1] + a[i] - b[i - 1] + b[i]) / 2
+    return area
 
 
 @cython.final